#326 feat: Move the converters into a subdirectory
